+++ examples/autopilot/assistant/basic/main.tf
@@ -1,3 +1,0 @@
-resource "twilio_autopilot_assistant" "assistant" {
-  friendly_name = "twilio-test"
-}