+++ examples/autopilot/assistant/stylesheet-and-defaults/outputs.tf
@@ -1,4 +1,0 @@
-output "assistant" {
-  description = "The Generated Autopilot Assistant"
-  value       = twilio_autopilot_assistant.assistant
-}