+++ examples/autopilot/field_type/main.tf
@@ -1,9 +1,0 @@
-resource "twilio_autopilot_assistant" "assistant" {
-  friendly_name = "test"
-}
-
-resource "twilio_autopilot_field_type" "field_type" {
-  assistant_sid = twilio_autopilot_assistant.assistant.sid
-  unique_name   = "test"
-  friendly_name = "test"
-}