+++ examples/autopilot/task/outputs.tf
@@ -1,9 +1,0 @@
-output "assistant" {
-  description = "The Generated Autopilot Assistant"
-  value       = twilio_autopilot_assistant.assistant
-}
-
-output "task" {
-  description = "The Generated Autopilot Task"
-  value       = twilio_autopilot_task.task
-}