+++ examples/studio/flow/with-conditional-widgets/outputs.tf
@@ -1,4 +1,0 @@
-output "flow" {
-  description = "The Generated Studio Flow"
-  value       = twilio_studio_flow.flow
-}