+++ twilio/common/twilio_client.go
@@ -3,7 +3,6 @@ package common
 import (
 	accounts "github.com/RJPearson94/twilio-sdk-go/service/accounts/v1"
 	api "github.com/RJPearson94/twilio-sdk-go/service/api/v2010"
-	autopilot "github.com/RJPearson94/twilio-sdk-go/service/autopilot/v1"
 	chat "github.com/RJPearson94/twilio-sdk-go/service/chat/v2"
 	conversations "github.com/RJPearson94/twilio-sdk-go/service/conversations/v1"
 	flex "github.com/RJPearson94/twilio-sdk-go/service/flex/v1"
@@ -24,7 +23,6 @@ type TwilioClient struct {
 
 	Accounts      *accounts.Accounts
 	API           *api.V2010
-	Autopilot     *autopilot.Autopilot
 	Chat          *chat.Chat
 	Conversations *conversations.Conversations
 	Flex          *flex.Flex