+++ twilio/internal/services/studio/tests/data_source_studio_flow_definition_test.go
@@ -18,7 +18,7 @@ func TestAccDataSourceTwilioStudioFlowDefinition_basic(t *testing.T) {
 			{
 				Config: testAccDataSourceTwilioStudioFlowDefinition_basic(),
 				Check: resource.ComposeTestCheckFunc(
-					resource.TestCheckResourceAttr(stateDataSourceName, "json", `{"description":"Bot flow for creating a Flex webchat task","flags":{"allow_concurrent_calls":true},"initial_state":"Trigger","states":[{"name":"Trigger","properties":{"offset":{"x":200,"y":0}},"transitions":[{"event":"incomingCall"},{"event":"incomingMessage","next":"SendToAutopilot"},{"event":"incomingParent"},{"event":"incomingRequest"}],"type":"trigger"},{"name":"SendMessageToAgent","properties":{"attributes":"{\"channelSid\":\"{{trigger.message.ChannelSid}}\",\"channelType\":\"{{trigger.message.ChannelAttributes.channel_type}}\",\"name\":\"{{trigger.message.ChannelAttributes.from}}\"}","channel":"TCaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa","offset":{"x":270,"y":540},"workflow":"WWaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"},"transitions":[{"event":"callComplete"},{"event":"callFailure"},{"event":"failedToEnqueue"}],"type":"send-to-flex"},{"name":"SendToAutopilot","properties":{"autopilot_assistant_sid":"UAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa","body":"{{trigger.message.Body}}","from":"{{flow.channel.address}}","offset":{"x":200,"y":240},"timeout":14400},"transitions":[{"event":"failure","next":"SendMessageToAgent"},{"event":"sessionEnded"},{"event":"timeout","next":"SendMessageToAgent"}],"type":"send-to-auto-pilot"}]}`),
+					resource.TestCheckResourceAttr(stateDataSourceName, "json", `{"description":"Flow for creating a Flex webchat task","flags":{"allow_concurrent_calls":true},"initial_state":"Trigger","states":[{"name":"Trigger","properties":{"offset":{"x":200,"y":0}},"transitions":[{"event":"incomingCall"},{"event":"incomingMessage","next":"SendMessageToAgent"},{"event":"incomingParent"},{"event":"incomingRequest"}],"type":"trigger"},{"name":"SendMessageToAgent","properties":{"attributes":"{\"channelSid\":\"{{trigger.message.ChannelSid}}\",\"channelType\":\"{{trigger.message.ChannelAttributes.channel_type}}\",\"name\":\"{{trigger.message.ChannelAttributes.from}}\"}","channel":"TCaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa","offset":{"x":270,"y":540},"workflow":"WWaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"},"transitions":[{"event":"callComplete"},{"event":"callFailure"},{"event":"failedToEnqueue"}],"type":"send-to-flex"}]}`),
 					helper.ValidateFlowDefinition(stateDataSourceName),
 				),
 			},
@@ -28,25 +28,6 @@ func TestAccDataSourceTwilioStudioFlowDefinition_basic(t *testing.T) {
 
 func testAccDataSourceTwilioStudioFlowDefinition_basic() string {
 	return `
-data "twilio_studio_flow_widget_send_to_autopilot" "send_to_autopilot" {
-  name = "SendToAutopilot"
-
-  offset {
-    x = 200
-    y = 240
-  }
-
-  transitions {
-    failure = data.twilio_studio_flow_widget_send_to_flex.send_to_flex.name
-    timeout = data.twilio_studio_flow_widget_send_to_flex.send_to_flex.name
-  }
-
-  autopilot_assistant_sid = "UAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
-  from                    = "{{flow.channel.address}}"
-  body                    = "{{trigger.message.Body}}"
-  timeout                 = 14400
-}
-
 data "twilio_studio_flow_widget_send_to_flex" "send_to_flex" {
   name = "SendMessageToAgent"
 
@@ -68,7 +49,7 @@ data "twilio_studio_flow_widget_trigger" "trigger" {
   name = "Trigger"
 
   transitions {
-    incoming_message = data.twilio_studio_flow_widget_send_to_autopilot.send_to_autopilot.name
+    incoming_message = data.twilio_studio_flow_widget_send_to_flex.send_to_flex.name
   }
 
   offset {
@@ -78,7 +59,7 @@ data "twilio_studio_flow_widget_trigger" "trigger" {
 }
 
 data "twilio_studio_flow_definition" "definition" {
-  description   = "Bot flow for creating a Flex webchat task"
+  description   = "Flow for creating a Flex webchat task"
   initial_state = data.twilio_studio_flow_widget_trigger.trigger.name
 
   flags {
@@ -91,10 +72,6 @@ data "twilio_studio_flow_definition" "definition" {
 
   states {
     json = data.twilio_studio_flow_widget_send_to_flex.send_to_flex.json
-  }
-
-  states {
-    json = data.twilio_studio_flow_widget_send_to_autopilot.send_to_autopilot.json
   }
 }
 `